corrected frontend service
