+++ softgroup/frontend/src/main/java/com/softgroup/frontend/configuration/RestApplicationConfiguration.java
@@ -1,4 +1,4 @@
-package com.softgroup.frontend.test.configuration;
+package com.softgroup.frontend.configuration;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
