+++ softgroup/frontend/src/main/java/com/softgroup/frontend/configuration/WebAppInitializer.java
@@ -1,4 +1,4 @@
-package com.softgroup.frontend.test.configuration;
+package com.softgroup.frontend.configuration;
 
 import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
 /**
