+++ softgroup/frontend/src/main/java/com/softgroup/frontend/configuration/WebConfiguration.java
@@ -1,0 +1,36 @@
+package com.softgroup.frontend.configuration;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.converter.HttpMessageConverter;
+import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+
+import java.util.List;
+
+/**
+ * Author: vadym
+ * Date: 01.03.17
+ * Time: 17:54
+ */
+@Configuration
+@EnableWebMvc
+@ComponentScan(basePackages = {
+        "com.softgroup.frontend"})
+public class WebConfiguration extends WebMvcConfigurerAdapter{
+    @Override
+    public void configureMessageConverters(List<HttpMessageConverter<?>> converters) {
+        converters.add(mappingJackson2HttpMessageConverter());
+    }
+
+    public MappingJackson2HttpMessageConverter mappingJackson2HttpMessageConverter() {
+        ObjectMapper mapper = new ObjectMapper();
+        mapper.disable(SerializationFeature.FAIL_ON_EMPTY_BEANS);
+        mapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+
+        return new MappingJackson2HttpMessageConverter(mapper);
+    }
+}
