+++ softgroup/frontend/src/main/java/com/softgroup/frontend/security/SecurityConfig.java
@@ -1,6 +1,5 @@
-package com.softgroup.frontend.test.security;
+package com.softgroup.frontend.security;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -14,11 +13,8 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
  */
 @Configuration
 @EnableWebSecurity
-@ComponentScan("com.softgroup.frontend.test.security")
+@ComponentScan("com.softgroup.frontend.security")
 public class SecurityConfig extends WebSecurityConfigurerAdapter {
-
-    @Autowired
-    Filter filter;
 
     @Override
     public void configure(WebSecurity web) throws Exception {
@@ -28,6 +24,6 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
     @Override
     protected void configure(HttpSecurity http) throws Exception {
         http.csrf().disable().authorizeRequests().anyRequest().authenticated();
-        http.addFilterBefore(filter, UsernamePasswordAuthenticationFilter.class);
+        http.addFilterBefore(new UserTokenFilter(), UsernamePasswordAuthenticationFilter.class);
     }
 }
