+++ softgroup/frontend/src/main/java/com/softgroup/frontend/security/TokenAuthenticationManager.java
@@ -1,4 +1,4 @@
-package com.softgroup.frontend.test.security;
+package com.softgroup.frontend.security;
 
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.core.Authentication;
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Component;
 /**
  * Created by vadym_polyanski on 20.03.17.
  */
-@Component
 public class TokenAuthenticationManager implements AuthenticationManager {
     @Override
     public Authentication authenticate(Authentication authentication) throws AuthenticationException {
