+++ softgroup/frontend/src/main/java/com/softgroup/frontend/security/UserTokenFilter.java
@@ -1,10 +1,7 @@
-package com.softgroup.frontend.test.security;
+package com.softgroup.frontend.security;
 
-import com.google.common.io.CharStreams;
-import com.softgroup.common.dao.api.entities.DeviceEntity;
-import com.softgroup.common.dao.api.entities.ProfileEntity;
 import com.softgroup.common.dao.impl.service.DeviceService;
-import com.softgroup.common.dao.impl.service.ProfileService;
+import com.softgroup.common.jwt.api.exception.SessionTokenException;
 import com.softgroup.common.jwt.impl.service.TokenService;
 import com.softgroup.common.protocol.RoutingData;
 import org.slf4j.Logger;
@@ -13,7 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.stereotype.Component;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
@@ -26,9 +22,8 @@ import java.io.IOException;
 /**
  * Created by vadym_polyanski on 20.03.17.
  */
-@Component
-public class Filter extends UsernamePasswordAuthenticationFilter {
-    static Logger log = LoggerFactory.getLogger(Filter.class);
+public class UserTokenFilter extends UsernamePasswordAuthenticationFilter {
+    static Logger log = LoggerFactory.getLogger(UserTokenFilter.class);
 
     @Autowired
     TokenAuthentication tokenAuthentication;
@@ -45,21 +40,19 @@ public class Filter extends UsernamePasswordAuthenticationFilter {
         log.info("SecurityFilter is working with request");
         HttpServletRequest request = (HttpServletRequest) req;
         String token = request.getHeader("x-token");
-        String deviceId = tokenService.getDeviceID(token);
 
-        DeviceEntity deviceEntity = deviceService.findDeviceEntityById(deviceId);
-
-
-        if (deviceEntity.getUpdateDateTime().equals(tokenService.getTimeOfCreation(token)) ) {
+        try {
+            tokenService.validateSessionToken(token);
             request.getSession().setAttribute("routing_data", createRoutingData(
                     request.getHeader("x-token")));
             SecurityContextHolder.getContext().setAuthentication(tokenAuthentication);
             chain.doFilter(req, res);
             log.info("Request status: OK");
-        } else {
+        } catch (SessionTokenException e) {
             ((HttpServletResponse) res).sendError(403);
             log.error("Request status: ERROR 403");
         }
+
     }
 
     private RoutingData createRoutingData(String token) {
